refactor: 로딩기능 추가 및 일부코드 변경
사용자의 정보가 다 불러와졌을때 로딩되게 바꾸었습니다.
navbar.js의 서버통신부분을 App.js에서 사용하게 바꾸었습니다.
