+++ README.md
@@ -22,12 +22,16 @@
 
 - 우측 상단에 codeit@codeit.com 은 코드잇서버에서 GET 하여 노출되고 있는것이며 만약 유저 데이터가 없다면 LOGIN 버튼이 노출이 됩니다. 또한 유저데이터가 없을시 Nav-bar 하단의 Section 도 노출이 되지않습니다. 여기서 유저 데이터의 존재유무는 userData?.id 를 사용하여 서버의 유저 객체에 id가 존재하나? 로 구분하였습니다. userData?.id[2]로 존재하지않는 id에 접근시 Section이 사라지며 Button이 노출되는것을 확인 가능합니다.
 
+  > 위의 기능은 수정되었습니다 App.js 컴포넌트에서 userData의 정보를 받아오게 만들었으며 props로 navbar.js에 전달해줍니다 만약 Userdata의 id가 존재하지 않을시에는 똑같이 로그인버튼을 보여주며 Main.js 또한 유저의 정보이기때문에 안 보입니다.
+
 - 서버데이터에 접근시 왠만하면 옵셔널 체이닝을 사용하여 서버데이터가 없으면 오류가아닌 Undefined가 반환되게 만들었습니다.
 
 - Main의 Card는 Hover시 이미지가 커지는 애니메이션을 가지고 있으며 각 Card또한 서버에서 가져온 데이터들입니다. 그리고 미션에는 없었지만 별도로 최신순 정렬이 되게 만들었습니다. 원래 Button을 만들어 최신순 , 이름순 정렬을 하고 싶었으나 다음 미션때 다 갈아엎을까봐
   그냥 서버에서 처음부터 최신순 정렬로 불러오게 만들었습니다.
 
 - 별도로 헤더의 Nav-bar에서 사용자 이메일 부분이랑 그 밑 섹션에 개발자/★즐겨찾기 부분에 a태그로 링크를 넣고싶었지만 미션에서 별 다른 링크를 제공하지않아서 그냥 내비뒀습니다. 코드 보실때 참고하시면 될 것 같습니다.
+
+- 추가 +) 로딩기능 재밌을것 같아서 구현해봤습니다.
 
 ## 2. 각 주차별 체크박스
 
@@ -52,6 +56,8 @@
 
 - 코드를 다 작성하고 문뜩 생각하여보니 Props를 한번도 사용을 안 하였습니다... 멘토님이 보시기엔 재가 작성한 코드에서 이 부분에서 Props 를 사용했으면 더 좋왔을거다 하는 부분이 있으셨나요 ? 아직 React에대한 저의 이해가 부족한것일수 있겠지만 주관적으로는 이번 미션에서 Props를 사용할 구간을 찾지 못하였습니다.
 - Props 를 부모컴포넌트가 자식컴포넌트의 속성을 조작할수있다. 라고만 일단 이해를 하고 있는데 솔직히 어디에 써야할지 잘 감이 안 옵니다. 이럴때 이럴때 사용 할 수 있다는 예시가 있을까요 ?
+
+> 수정 ) App.js 에서 props를 사용하였습니다 ! 이번 페이지는 결국 유저의 로그인 데이터를 불러오는것에 따라 정보가 표시되고 안 표시되고를 구현하는게 중요하다 생각하여 App.js 에서 서버에서 유저데이터를 Get하는것을 다루고 존재여부에따라 다른 컴포넌트가 표시될지 안 표시될지를 만들었습니다 !
 
 ### 2. AI에 의존하다
