+++ src/component/loding/Loding.css
@@ -1,0 +1,19 @@
+.background {
+  position: absolute;
+  width: 100vw;
+  height: 100vh;
+  top: 0;
+  left: 0;
+  background: #00000070;
+  z-index: 999;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.lodingText {
+  text-align: center;
+  font-size: 20px;
+  font-weight: 600;
+}
