+++ src/component/shared/header/Navbar.js
@@ -1,33 +1,7 @@
-import { useEffect, useState } from "react";
 import "./Header.css";
 
-import HeaderFoloderSection from "./HeaderFolderSection";
-
-function Navbar() {
-  // 유저 데이터를 담을 State
-  const [userData, setUserData] = useState({});
-
-  // 사이드 이펙트 처리 & 데이터 GET
-  useEffect(() => {
-    async function getServerData() {
-      try {
-        const res = await fetch(
-          "https://bootcamp-api.codeit.kr/api/sample/user"
-        );
-        const data = await res.json();
-        setUserData(data);
-      } catch (e) {
-        console.log("에러 :" + e);
-      }
-    }
-
-    getServerData();
-  }, []);
-
-  //이름 추상화
+function Navbar({ userData }) {
   const { email, profileImageSource } = userData;
-  // 데이터 확인용
-  //console.log(userData);
 
   return (
     <>
@@ -55,7 +29,6 @@ function Navbar() {
           )}
         </div>
       </nav>
-      {userData?.id ? <HeaderFoloderSection /> : ""}
     </>
   );
 }
